pyproject.toml
[project]
name = "tgbotanna"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.4",
    "eth-account>=0.13.7",
    "openai>=2.24.0",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.6",
    "solana>=0.36.11",
    "solders>=0.27.1",
    "web3>=7.14.1",
]
